Move QueuedStream so it can be used on mobile
