+++ MyPal.Console/Program.cs
@@ -11,7 +11,7 @@ Console.ReadLine();
 class Microphone : IMicrophone
 {
     readonly WaveInEvent _waveInEvent;
-    BufferStream _stream = new();
+    QueuedStream _stream = new();
 
     public Microphone()
     {
@@ -27,44 +27,6 @@ class Microphone : IMicrophone
     }
 
     public Stream GetAudio() => _stream;
-}
-
-class BufferStream : Stream
-{
-    ConcurrentQueue<byte[]> _buffers = new();
-
-    public override bool CanRead => true;
-
-    public override bool CanSeek => false;
-
-    public override bool CanWrite => false;
-
-    public override long Length => throw new NotImplementedException();
-
-    public override long Position { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
-    public override void Flush() => throw new NotImplementedException();
-
-    public void Enqueue(byte[] buffer) => _buffers.Enqueue(buffer);
-
-    public override int Read(byte[] buffer, int offset, int count)
-    {
-        byte[]? bytes;
-        while (!_buffers.TryDequeue(out bytes))
-        {
-            Thread.Sleep(100);
-        }
-
-        int length = Math.Min(bytes.Length, count);
-        Array.Copy(bytes, 0, buffer, offset, length);
-        return length;
-    }
-
-    public override long Seek(long offset, SeekOrigin origin) => throw new NotImplementedException();
-
-    public override void SetLength(long value) => throw new NotImplementedException();
-
-    public override void Write(byte[] buffer, int offset, int count) => throw new NotImplementedException();
 }
 
 class Speaker : ISpeaker, IDisposable
